+++ src/frequenz/sdk/timeseries/_formula_engine/_formula_generators/__init__.py
@@ -4,7 +4,6 @@
 """Generators for formulas from component graphs."""
 
 from ._battery_power_formula import BatteryPowerFormula
-from ._battery_soc_formula import BatterySoCFormula
 from ._ev_charger_current_formula import EVChargerCurrentFormula
 from ._ev_charger_power_formula import EVChargerPowerFormula
 from ._formula_generator import (
@@ -28,7 +27,6 @@ __all__ = [
     #
     "GridPowerFormula",
     "BatteryPowerFormula",
-    "BatterySoCFormula",
     "EVChargerPowerFormula",
     "PVPowerFormula",
     #